Templated library and split up code
Everything is templated off the point class and the number of
dimensions.

Stripped down the Point2 class to represent the minimum interface
required for the kdtree to work. Requires an array operator
interface for getting component values, the ability to create one
from an appropriately lengthed array and a base_type typedef which
reveals the component data types.

Added a Measurer class to do the calculation of distances between two
points so that a substraction operator and some kind of lengthSq()
operator need not be provided by the user.

TreeFactory is not templated so that it can be used as a member of a
class without templating that class. Same for BoundsFactory though users
aren't expected to use that beyond setting up the TreeFactory.